+++ src/Graylog.Target/IConverter.cs
@@ -11,17 +11,16 @@ using NLog;
 namespace Graylog.Target
 {
 	/// <summary>
-	/// Конвертер событий логирования в объекты JSON.
+	/// Converter to GELF-style JSON.
 	/// </summary>
 	public interface IConverter
 	{
 		/// <summary>
-		/// Возвращает объект созданный из сообщения.
+		/// Create JSON object from <see cref="LogEventInfo"/>.
 		/// </summary>
-		/// <param name="logEventInfo">Информация о сообщении.</param>
-		/// <param name="facility">Значение настройки facility.</param>
-		/// <param name="includeMdlcProperties">If <c>true</c> include <see cref="MappedDiagnosticsLogicalContext"/> properties into message.</param>
-		/// <returns>Объект для сериализации.</returns>
-		JObject GetGelfJson(LogEventInfo logEventInfo, string facility, bool includeMdlcProperties = false);
+		/// <param name="logEventInfo"><see cref="LogEventInfo"/>.</param>
+		/// <param name="options">Convert options.</param>
+		/// <returns>GELF-style JSON.</returns>
+		JObject GetGelfJson(LogEventInfo logEventInfo, IConvertOptions options);
 	}
 }